requirements.txt
gensim==4.1.2
nltk==3.6.5
numpy==1.21.4
ordered-set==4.0.2
pandas==1.3.4
PyYAML==6.0
regex==2021.11.10
scikit-learn==1.0.1
sklearn==0.0
tensorflow==2.7.0
tensorflow-datasets==4.4.0
